% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE90102_01.R
\docType{data}
\name{GSE90102_01}
\alias{GSE90102_01}
\title{Subset of SNPs from meta-analysis of Crohn's disease GWAS study.}
\format{
A list with several components:
\describe{
\item{rawp}{P-value for association between SNP and Crohn's disease.}
\item{MAF}{Minor Allele Frequency (MAF) of SNP.}
}
}
\usage{
data(GSE90102_01)
}
\value{
Object of class list.
}
\description{
A dataset containing 607,720 SNPs and results for their associations with Crohn's disease.
}
\keyword{datasets}
