% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu-data-tutorial.R
\docType{data}
\name{simulation_data}
\alias{simulation_data}
\title{List of simulation datasets}
\format{
A list of 390 datasets, for each of them:
\describe{
\item{res_FDR_power}{A dataframe containing the FDR and power of each procedure.}
\item{res_Subpower}{A dataframe containing the subpower of each procedure.}
}
}
\usage{
data(simulation_data)
}
\value{
Object of class list.
}
\description{
A list containing 390 results from the simulation study.
Each dataset contains two dataframes: res_FDR_power and res_Subpower.
}
