% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBHa.R
\name{wBHa}
\alias{wBHa}
\title{wBHa Procedure}
\usage{
wBHa(pvalues, covariates, alpha = 0.05, K = 60)
}
\arguments{
\item{pvalues}{Numeric vector of pvalues.}

\item{covariates}{Numeric vector of covariates independent under the H0 of the pvalue.}

\item{alpha}{Numeric in \code{[}0,1\code{]}, significance level of the multiple testing procedure. Default \code{alpha=0.05}.}

\item{K}{Integer, number of bootstrap samples. Default \code{K=60}.}
}
\value{
\describe{
\item{\code{final_a_opt}}{Numeric value of the parameter a.}
\item{\code{adjusted_covariates}}{Numeric vector of adjusted covariates.}
\item{\code{adjusted_pvalues}}{Numeric vector of adjusted pvalues.}
}
}
\description{
wBHa is an R package that implements a covariate multiple testing procedure specifically adapted in a Genome-Wide Association Studies context.
wBHa allows better detection of rare variants (which where difficult to detect since the current existing procedures are not powerful enough) by better integrating external information while allowing the optimization of the overall power.
}
\examples{
pvalues<-runif(100)
covariates<-runif(100,1,2)
wBHa(pvalues,covariates,alpha=0.05,K=60)

}
\author{
Ludivine Obry
}
